requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
greenlet==3.3.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.13.0.90
opencv-python-headless==4.11.0.86
oracledb==3.4.1
packaging==26.0
pillow==10.2.0
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
sympy==1.14.0
torch==2.1.2
torchvision==0.16.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.7
ultralytics-thop==2.0.18
urllib3==2.6.3
